requirements.txt
tensorflow>=2.8.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0
nltk>=3.6.0
tqdm>=4.62.0
grpcio>=1.24.3,<2.0.0
seaborn>=0.12.0
kagglehub>=0.1.5
python-dotenv>=1.0.0
contractions>=0.1.72
html5lib>=1.1
transformers>=4.30.0
sentencepiece>=0.1.99 